refactor: bottom bar 디렉토리 정리 #3
